+++ NEWS.md
@@ -1,4 +1,4 @@
-# babelgene (development version)
+# babelgene 22.9
 
 * Updated the database.
 * Added a check for NAs in input genes (#3).